# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Week13 API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/post": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get post list",
                "description": "Get post list",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "Filter by category_id",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "Filter by title",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "Filter by content",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "author",
                        "in": "formData",
                        "description": "Filter by author",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post list",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "properties": {
                                        "post": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Post"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update post",
                "description": "Update post",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "Filter by category_id",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "Filter by title",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "Filter by content",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "author",
                        "in": "formData",
                        "description": "Filter by author",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post was updated successfully",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "properties": {
                                        "post": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Post"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing required fields"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create new post",
                "description": "Create new post",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "Filter by category_id",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "Filter by title",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "Filter by content",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "author",
                        "in": "formData",
                        "description": "Filter by author",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post was stored successfully",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "properties": {
                                        "post": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Post"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Missing required fields"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete post",
                "description": "Delete post",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "formData",
                        "description": "Filter by category_id",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "Filter by title",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "Filter by content",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "author",
                        "in": "formData",
                        "description": "Filter by author",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post was deleted successfully",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "properties": {
                                        "post": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Post"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Post": {
            "required": [
                "label",
                "category_id",
                "title"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "category_id": {
                    "description": "category_id",
                    "type": "integer"
                },
                "title": {
                    "description": "title",
                    "type": "string"
                },
                "content": {
                    "description": "content",
                    "type": "string"
                },
                "author": {
                    "description": "author",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        }
    }
}